feat(video-player): migrate to videojs v10
