--- packages/video-player-v10/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-dist
-
-build
-*.tsbuildinfo
