--- packages/video-player-v10/.npmignore
@@ -1,12 +1,0 @@
-.gitignore
-.gitmodules
-src
-node_modules
-.prettierignore
-.prettierrc
-index.html
-postcss.config.js
-vite.config.ts
-demo
-pnpm-lock.yaml
-external-projects
