--- packages/video-player-v10/.prettierignore
@@ -1,6 +1,0 @@
-node_modules
-dist
-.vscode
-demo
-external-projects
-pnpm-lock.yaml
