--- packages/video-player-v10/demo/main.ts
@@ -1,5 +1,0 @@
-import { createApp } from "vue";
-import App from "./App.vue"
-
-createApp(App)
-    .mount("#app")
