--- packages/video-player-v10/index.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <title>BrunstadTV Player test</title>
-    </head>
-    <body>
-        <div id="app"></div>
-        <script type="module" src="/demo/main.ts"></script>
-    </body>
-</html>
