--- packages/video-player-v10/package.json
@@ -1,47 +1,0 @@
-{
-    "name": "bccm-video-player-v10",
-    "version": "0.0.0-dev",
-    "type": "commonjs",
-    "types": "build/types/index.d.ts",
-    "files": [
-        "build"
-    ],
-    "exports": {
-        ".": {
-            "types": "./build/types/index.d.ts",
-            "import": "./build/bccm-video-player.mjs",
-            "require": "./build/bccm-video-player.umd.js"
-        },
-        "./css": "./build/bccm-video-player.css"
-    },
-    "scripts": {
-        "dev": "vite",
-        "types": "tsc --emitDeclarationOnly",
-        "build": "tsc --noEmit && vite build && pnpm types",
-        "preview": "vite preview",
-        "prettier": "prettier --write ."
-    },
-    "dependencies": {
-        "@videojs/core": "10.0.0-beta.23",
-        "@videojs/element": "10.0.0-beta.23",
-        "@videojs/html": "10.0.0-beta.23",
-        "graphql": "^16.13.2",
-        "graphql-request": "^7.4.0",
-        "hls.js": "^1.6.16",
-        "npaw-plugin-adapters": "^7.3.29",
-        "npaw-plugin-nwf": "^7.3.29"
-    },
-    "devDependencies": {
-        "@types/node": "^25.6.0",
-        "@vitejs/plugin-vue": "^6.0.6",
-        "@vue/tsconfig": "^0.9.1",
-        "postcss": "^8.5.14",
-        "prettier": "^3.8.3",
-        "sass": "^1.99.0",
-        "typescript": "^6.0.3",
-        "vite": "^8.0.10",
-        "vue": "^3.5.33",
-        "vue-tsc": "^3.2.8"
-    },
-    "packageManager": "pnpm@9.14.4"
-}
