--- packages/video-player-v10/postcss.config.js
@@ -1,1 +1,0 @@
-module.exports = {}
