--- packages/video-player-v10/src/btv-player/api/index.ts
@@ -1,2 +1,0 @@
-export * from "./episode"
-export * from "./client"
