--- packages/video-player-v10/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./video-player"
-export * from "./btv-player"
