--- packages/video-player-v10/src/video-player/utils/debounce.js
@@ -1,9 +1,0 @@
-export default function (a, b, c, d) {
-    return function () {
-        d = this
-        clearTimeout(c)
-        c = setTimeout(function () {
-            a.apply(d, arguments)
-        }, b || 100)
-    }
-}
