--- packages/video-player-v10/src/video-player/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./localeMapper"
