--- packages/video-player-v10/src/vite-env.d.ts
@@ -1,7 +1,0 @@
-/// <reference types="vite/client" />
-
-declare module "*.vue" {
-    import type { DefineComponent } from "vue"
-    const component: DefineComponent<{}, {}, any>
-    export default component
-}
