--- packages/video-player-v10/tsconfig.json
@@ -1,23 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "noImplicitAny": false,
-        "allowJs": true,
-        "useDefineForClassFields": true,
-        "module": "ESNext",
-        "moduleResolution": "Bundler",
-        "strict": true,
-        "jsx": "preserve",
-        "rootDir": "src",
-        "outDir": "build/types",
-        "sourceMap": true,
-        "resolveJsonModule": true,
-        "isolatedModules": true,
-        "esModuleInterop": true,
-        "lib": ["ESNext", "DOM"],
-        "skipLibCheck": true,
-        "declaration": true,
-    },
-    "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue"],
-    "references": [{ "path": "./tsconfig.node.json" }],
-}
