--- packages/video-player-v10/vite.config.ts
@@ -1,20 +1,0 @@
-import { defineConfig } from "vite"
-import vue from "@vitejs/plugin-vue"
-import path from "path"
-
-// https://vitejs.dev/config/
-export default defineConfig({
-    plugins: [vue()],
-    resolve: {
-        alias: {
-            "@": path.resolve(__dirname, "src"),
-        },
-    },
-    build: {
-        lib: {
-            entry: path.resolve(__dirname, "src"),
-            name: "btv-video",
-        },
-        outDir: "build",
-    },
-})
