+++ packages/video-player/demo/tsconfig.json
