+++ packages/video-player/src/video-player/components/dismiss-controls-button.ts
