+++ packages/video-player/src/video-player/components/playback-rate-picker.ts
