+++ packages/video-player/src/video-player/components/quality-picker.ts
