+++ packages/video-player/src/video-player/npaw.ts
@@ -1,6 +1,12 @@
 import NpawPlugin from "npaw-plugin-nwf"
-import NpawAdapters from "npaw-plugin-adapters"
+import * as NpawAdapters from "npaw-plugin-adapters"
+import Hls from "hls.js"
 import type { Player } from "./index"
+
+// NPAW's HlsjsAdapter uses bare `Hls.Events.X` references and expects `Hls`
+// to be a runtime global (legacy <script>-tag assumption). Ensure it's
+// reachable before any adapter registration.
+;(globalThis as { Hls?: unknown }).Hls = Hls
 
 export interface NPAWOptions {
     enabled?: boolean
@@ -24,7 +30,7 @@ export interface NPAWOptions {
     }
 }
 
-function toConfig(options: NPAWOptions) {
+function toConfig(options: NPAWOptions): Record<string, unknown> {
     const md = options.tracking.metadata
     return {
         "content.isLive": options.tracking.isLive === true,
@@ -36,38 +42,141 @@ function toConfig(options: NPAWOptions) {
             ? `${md.seasonId} - ${md.seasonTitle}`
             : undefined,
         "content.episodeTitle": md.episodeTitle,
-        obfuscateIp: true,
+        "user.obfuscateIp": true,
         "user.name": options.tracking.userId,
         "app.name": options.appName,
-        "app.releaseVersion": "", //RELEASE_VERSION,
+        "app.releaseVersion": "",
         "parse.manifest": true,
-        "extraparam.1": options.tracking.sessionId,
-        "extraparam.2": options.tracking.ageGroup,
+        "content.customDimension.1": options.tracking.sessionId,
+        "content.customDimension.2": options.tracking.ageGroup,
         ...md.overrides,
     }
 }
 
-export function enableNPAW(player: Player, options: NPAWOptions) {
+// NPAW ships an HlsjsAdapter that consumes the hls.js engine instance
+// directly — no custom adapter needed. v10's <hls-video> exposes `.engine`
+// once the manifest starts loading; we wait for it before registering.
+export function enableNPAW(
+    player: Player,
+    options: NPAWOptions,
+    signal?: AbortSignal
+): void {
     if (!options.accountCode) {
         console.warn(
             "NPAW was not enabled because options.npaw.accountCode is invalid."
         )
         return
     }
+
     const npaw = new NpawPlugin(options.accountCode)
-    const defaults = toConfig(options)
+    npaw.setAnalyticsOptions(toConfig(options))
+    ;(player as Player & { _npaw?: NpawPlugin })._npaw = npaw
 
-    npaw.setAnalyticsOptions(defaults)
-    npaw.registerAdapterFromClass(player, NpawAdapters.video.Videojs)
+    waitForEngine(player.mediaEl, (engine) => {
+        npaw.registerAdapterFromClass(
+            engine,
+            (NpawAdapters as { HlsjsAdapter: unknown }).HlsjsAdapter
+        )
 
-    ;(player as any)._npaw = npaw
+        // NPAW's adapter calls `checkExistsPlayer()` on every fire and walks
+        // parentNode from `engine.media` to verify it's connected to document.
+        // v10's <hls-video> nests the inner video inside a shadow DOM, so the
+        // walk hits the shadow boundary instead of document and returns false
+        // — NPAW then refuses to fire any events ("Cannot fire start event
+        // because player not exists on the document"). The element IS on the
+        // page (via the shadow host), so override the check.
+        const adapter = npaw.getAdapter() as
+            | { checkExistsPlayer?: () => boolean }
+            | undefined
+        if (adapter) adapter.checkExistsPlayer = () => true
+
+        // VideoJsAdapter (used in v8) reported text-track changes natively.
+        // HlsjsAdapter doesn't, so wire equivalent custom events to keep
+        // BTV's audio/subtitle engagement metrics intact post-cutover.
+        wireTrackChangeEvents(player.mediaEl, engine, npaw, signal)
+    })
 }
 
-export function setOptions(player: Player, options: NPAWOptions) {
-    const npaw = (player as any)._npaw
-    if (!npaw) {
+export function setOptions(player: Player, options: NPAWOptions): void {
+    const npaw = (player as Player & { _npaw?: NpawPlugin })._npaw
+    if (!npaw) return
+    npaw.setAnalyticsOptions(toConfig(options))
+}
+
+type HlsAudioTrack = { lang?: string; name?: string }
+type HlsEngine = {
+    audioTracks?: HlsAudioTrack[]
+    audioTrack: number
+    on(event: string, cb: () => void): void
+    off(event: string, cb: () => void): void
+}
+
+function wireTrackChangeEvents(
+    mediaEl: HTMLElement,
+    engine: unknown,
+    npaw: NpawPlugin,
+    signal: AbortSignal | undefined
+): void {
+    const adapter = npaw.getAdapter() as
+        | { fireEvent(name: string, dimensions?: object): void }
+        | undefined
+    if (!adapter?.fireEvent) return
+
+    // Subtitles: v10's hls.js promotes subtitle tracks onto media.textTracks,
+    // so DOM events cover both native HLS and MSE paths.
+    const tracks = (mediaEl as HTMLVideoElement).textTracks
+    const onSubChange = () => {
+        const active = Array.from(tracks).find(
+            (t) =>
+                t.mode === "showing" &&
+                (t.kind === "subtitles" || t.kind === "captions")
+        )
+        adapter.fireEvent("subtitleChange", {
+            language: active?.language ?? "off",
+            label: active?.label ?? "off",
+        })
+    }
+    tracks.addEventListener("change", onSubChange)
+    signal?.addEventListener("abort", () =>
+        tracks.removeEventListener("change", onSubChange)
+    )
+
+    // Audio tracks: hls.js owns the list; subscribe to its switch events.
+    const eng = engine as HlsEngine
+    if (typeof eng.on === "function") {
+        const onAudioChange = () => {
+            const active = eng.audioTracks?.[eng.audioTrack]
+            adapter.fireEvent("audioChange", {
+                language: active?.lang ?? "",
+                label: active?.name ?? "",
+            })
+        }
+        eng.on("hlsAudioTrackSwitched", onAudioChange)
+        signal?.addEventListener("abort", () =>
+            eng.off("hlsAudioTrackSwitched", onAudioChange)
+        )
+    }
+}
+
+// hls.js engine becomes available asynchronously after src is set. Poll
+// until present, then invoke. Returns a teardown so callers can cancel
+// (currently unused — registration is fire-and-forget for the player's
+// lifetime; NPAW handles its own cleanup when the engine is destroyed).
+function waitForEngine(
+    mediaEl: HTMLElement,
+    cb: (engine: unknown) => void
+): void {
+    const get = () => (mediaEl as { engine?: unknown | null }).engine ?? null
+    const initial = get()
+    if (initial) {
+        cb(initial)
         return
     }
-
-    Object.assign(npaw.options, toConfig(options))
+    const id = setInterval(() => {
+        const engine = get()
+        if (engine) {
+            clearInterval(id)
+            cb(engine)
+        }
+    }, 250)
 }
