--- packages/video-player/src/video-player/skin/_small-screen-ui.scss
@@ -1,67 +1,0 @@
-@use "variables";
-
-.vjs-smallscreen-ui {
-    .vjs-control-bar {
-        height: 100%;
-        justify-content: flex-end;
-        align-content: baseline;
-        flex-wrap: wrap;
-        padding-top: 0.5em;
-
-        &:before {
-            height: 100%;
-            margin-left: 0;
-            margin-right: variables.$side-padding * -1;
-            background: rgba(0, 0, 0, 0.4);
-            box-shadow: inset 0 0 50em rgb(0 0 0 / 40%);
-        }
-        .vjs-control {
-            height: 3em;
-        }
-        .vjs-seek-to-live-control {
-            align-items: flex-start;
-        }
-        .vjs-play-control {
-            position: absolute;
-            left: 40%;
-            top: 40%;
-            min-width: 3em;
-            min-height: 3em;
-            width: 20%;
-            height: 20%;
-            > .vjs-icon-placeholder {
-                font-size: 2.5em;
-                &:before {
-                    height: 100% !important;
-                    /* top: 50% !important; */
-                    display: flex;
-                    align-items: center;
-                    justify-content: center;
-                }
-            }
-        }
-        .vjs-time-divider {
-            display: none;
-        }
-        .vjs-current-time,
-        .vjs-duration,
-        .vjs-seek-to-live-control {
-            position: absolute;
-            bottom: 2em;
-        }
-        .vjs-menu {
-            bottom: auto;
-            top: 3em;
-            .vjs-menu-content {
-                bottom: auto;
-                top: 0;
-            }
-        }
-        .vjs-current-time {
-            left: variables.$side-padding;
-        }
-        .vjs-progress-control {
-            bottom: 0.2em;
-        }
-    }
-}
