--- packages/video-player/src/video-player/skin/_variables.scss
@@ -1,2 +1,0 @@
-$other-fonts: Arial, Helvetica, sans-serif;
-$side-padding: 10px;
