+++ packages/video-player/src/video-player/skin/skin.css
