+++ packages/video-player/src/video-player/utils/index.ts
@@ -1,2 +1,1 @@
 export * from "./localeMapper"
-export * from "./castLoader"
