+++ packages/video-player/tsconfig.json
@@ -8,7 +8,8 @@
         "moduleResolution": "Bundler",
         "strict": true,
         "jsx": "preserve",
-        "outDir": "build",
+        "rootDir": "src",
+        "outDir": "build/types",
         "sourceMap": true,
         "resolveJsonModule": true,
         "isolatedModules": true,
